requirements.txt
Flask==2.0.1
requests==2.25.1
beautifulsoup4==4.9.3
jinja2==3.0.1
pdfkit==0.6.1
PyYAML==5.4.1
urllib3==1.26.5